Changed the functionality (post-processing folder). I added it into the slicer, and split it up into pre-processing and post-processing
