+++ examples/1_planar_slicing_simple/example_1_planar_slicing_simple.py
@@ -1,16 +1,14 @@
 import os
 from compas.datastructures import Mesh
-from compas.geometry import Point, Frame
+from compas.geometry import Point
 
 from compas_slicer.utilities import save_to_json
 from compas_slicer.slicers import PlanarSlicer
-from compas_slicer.post_processing import generate_brim
-from compas_slicer.post_processing import spiralize_contours
-from compas_slicer.post_processing import seams_align
-from compas_slicer.post_processing import seams_smooth, unify_paths_orientation
+from compas_slicer.slicers.post_processing import seams_smooth
 from compas_slicer.print_organization import PrintOrganizer
 from compas_viewers.objectviewer import ObjectViewer
-from compas_slicer.post_processing import move_mesh_to_point, simplify_paths_rdp
+from compas_slicer.slicers.post_processing import simplify_paths_rdp, generate_brim
+from compas_slicer.slicers.pre_processing import move_mesh_to_point
 import time
 
 ######################## Logging
@@ -39,7 +37,7 @@ def main():
     slicer.slice_model()
 
     ### --- Generate brim
-    # generate_brim(slicer, layer_width=3.0, number_of_brim_paths=3)
+    generate_brim(slicer, layer_width=3.0, number_of_brim_paths=3)
 
     ### --- Simplify the printpaths by removing points with a certain threshold
     # change the threshold value to remove more or less points
