+++ examples/2_planar_slicing_advanced/example_2_planar_slicing_advanced.py
@@ -1,17 +1,14 @@
 import os
 from compas.datastructures import Mesh
-from compas.geometry import Point, Frame
+from compas.geometry import Point
 
 from compas_slicer.utilities import save_to_json
 from compas_slicer.slicers import PlanarSlicer
-from compas_slicer.post_processing import generate_brim
-from compas_slicer.post_processing import spiralize_contours
-from compas_slicer.post_processing import seams_align
-from compas_slicer.post_processing import seams_smooth, unify_paths_orientation
+from compas_slicer.slicers.post_processing import generate_brim
 from compas_slicer.print_organization import PrintOrganizer
 from compas_viewers.objectviewer import ObjectViewer
-from compas_slicer.post_processing import move_mesh_to_point, simplify_paths_rdp
-import time
+from compas_slicer.slicers.post_processing import simplify_paths_rdp
+from compas_slicer.slicers.pre_processing import move_mesh_to_point
 
 ######################## Logging
 import logging
