+++ src/compas_slicer/__init__.py
@@ -62,7 +62,6 @@ except Exception:
 from .geometry import *  # noqa: F401 E402 F403
 from .slicers import *  # noqa: F401 E402 F403
 from .print_organization import *  # noqa: F401 E402 F403
-from .post_processing import *  # noqa: F401 E402 F403
 from .utilities import *  # noqa: F401 E402 F403
