+++ src/compas_slicer/slicers/__init__.py
@@ -34,5 +34,7 @@ from .planar_slicer import *  # noqa: F401 E402 F403
 from .curved_slicer import *  # noqa: F401 E402 F403
 from .planar_slicing import *  # noqa: F401 E402 F403
 from .curved_slicing import *  # noqa: F401 E402 F403
+from .post_processing import *  # noqa: F401 E402 F403
+from .pre_processing import *  # noqa: F401 E402 F403
 
 __all__ = [name for name in dir() if not name.startswith('_')]
