+++ src/compas_slicer/slicers/post_processing/__init__.py
@@ -23,17 +23,6 @@ Sorting
     seams_smooth
     sort_per_segment
 
-Positioning
-===========
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    center_on_build_platform
-    move_mesh_to_point
-    get_mid_pt_base
-
 Additional
 ==========
 
@@ -60,11 +49,6 @@ from .sort_paths_per_vertical_segment import *  # noqa: F401 E402 F403
 
 #  Orienting
 from .unify_paths_orientation import *  # noqa: F401 E402 F403
-
-#  Positioning
-from .center_on_build_platform import *  # noqa: F401 E402 F403
-from .move_mesh_to_point import *  # noqa: F401 E402 F403
-from .get_mid_pt_base import *  # noqa: F401 E402 F403
 
 #  Additional
 from .generate_brim import *  # noqa: F401 E402 F403
