+++ src/compas_slicer/slicers/post_processing/seams_smooth.py
