+++ src/compas_slicer/slicers/post_processing/simplify_paths_rdp.py
