+++ src/compas_slicer/slicers/post_processing/sort_paths_per_vertical_segment.py
