+++ src/compas_slicer/slicers/post_processing/spiralize_contours.py
