+++ src/compas_slicer/slicers/pre_processing/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+#  Positioning
+from .center_on_build_platform import *  # noqa: F401 E402 F403
+from .move_mesh_to_point import *  # noqa: F401 E402 F403
+from .get_mid_pt_base import *  # noqa: F401 E402 F403
