+++ src/compas_slicer/slicers/pre_processing/get_mid_pt_base.py
